testing pipeline

+++ src/main/java/se/anton/userregistryspring/app/dto/appuser/AppUserRequest.java
@@ -3,12 +3,14 @@ package se.anton.userregistryspring.app.dto.appuser;
 
 //todo add validation
 
+import java.math.BigDecimal;
+
 public record AppUserRequest(
         String email,
         String password,
         String firstName,
         String lastName,
-        Float funds
+        BigDecimal funds
 ) {
 
 }
