+++ src/main/java/se/anton/userregistryspring/app/dto/appuser/AppUserResponse.java
@@ -1,11 +1,13 @@
 package se.anton.userregistryspring.app.dto.appuser;
 
+import java.math.BigDecimal;
+
 public record AppUserResponse(
         String id,
         String email,
         String firstName,
         String lastName,
-        Float funds,
+        BigDecimal funds,
         String roles
 ) {
 }
